Made some changes to the project to prepare for the Asset Store submission: 1. Created a new variable in uScriptConfig.cs called "CommercialBuild". Set to true when building an Asset Store build so it disabled date and Unity version checking DRM code. 2. Fixed vertical offset of text below sockets. 3. Added a horizontal offset variable to uScriptConfig.cs called "SocketValueTextHorizontalOffset" to allow for horizontal adjustment (offset by 2 pixels makes it look more centered with the italic font used). 4. Updated Build Name to say "Retail Beta 1". 5. Fixed error in OS X when trying to copy Gizmos files by checking File.Exists first.
git-svn-id: https://www.detoxstudios.com/projectdepot/Development/Unity@1171 b191eb17-2330-428c-a13f-159e02ee8e09

+++ uScript_Dev/Assets/uScript/uScriptEditor/Editor/uScript.cs
@@ -30,12 +30,13 @@ public class uScript : EditorWindow
    private List<ZoomLink> m_ZoomLinks = null;
 
    //format is MAJOR.MINOR.FOURDIGITSVNCOMMITNUMBER
+   public bool CommercialBuild = false; // Set to false for free forum builds so it will use ExpireDate.
    public string uScriptBuild { get { return "0.9.1160"; } }
-   static string BetaVersion { get { return "Beta 11"; } }
+   static string BetaVersion { get { return "Retail Beta 1"; } }
    public string LastUnityBuild { get { return "3.3"; } }
    public string CurrentUnityBuild { get { return "3.4"; } }
    public string BetaUnityBuild { get { return "3.5"; } }
-   public DateTime ExpireDate { get { return new DateTime(2011, 9, 15); } }
+   public DateTime ExpireDate { get { return new DateTime(2011, 10, 31); } }
    public int EULAVersion { get { return 20110824; } }
 
    public enum MouseRegion
@@ -375,7 +376,10 @@ Should you have any questions concerning this EULA, or if you desire to contact 
       
       foreach ( System.IO.FileInfo file in directory.GetFiles( ) )
       {
-         System.IO.File.Copy( file.FullName, gizmos + "/" + file.Name, false );
+		 if (!System.IO.File.Exists( gizmos + "/" + file.Name ))
+			{
+				System.IO.File.Copy( file.FullName, gizmos + "/" + file.Name, false );
+			}
       }
    }
 
@@ -570,25 +574,29 @@ Should you have any questions concerning this EULA, or if you desire to contact 
       {
          if (null == m_ScriptEditorCtrl)
          {
-            //if ( Application.unityVersion == RequiredUnityBuild || Application.unityVersion == RequiredUnityBetaBuild || Application.unityVersion == RequiredUnityBetaBuildPrevious )
-            if (Application.unityVersion.Contains(LastUnityBuild) || Application.unityVersion.Contains(CurrentUnityBuild) || Application.unityVersion.Contains(BetaUnityBuild))
-            {
-            }
-            else
-            {
-               uScriptDebug.Log(BetaVersion + " (" + uScriptBuild + ") " + "will not work with Unity version " + Application.unityVersion + ".", uScriptDebug.Type.Error);
-               return;
-            }
-
-            if (DateTime.Now > ExpireDate)
-            {
-               uScriptDebug.Log(BetaVersion + " (" + uScriptBuild + ") " + "has expired.\n", uScriptDebug.Type.Error);
-               return;
-            }
-            else
-            {
-               uScriptDebug.Log(BetaVersion + " (" + uScriptBuild + ") " + "will expire in " + (ExpireDate - DateTime.Now).Days + " days.", uScriptDebug.Type.Message);
-            }
+			if (!CommercialBuild) // See if expiration date and build cap should be used. Not needed for commercial version.
+		    {
+	            //if ( Application.unityVersion == RequiredUnityBuild || Application.unityVersion == RequiredUnityBetaBuild || Application.unityVersion == RequiredUnityBetaBuildPrevious )
+	            if (Application.unityVersion.Contains(LastUnityBuild) || Application.unityVersion.Contains(CurrentUnityBuild) || Application.unityVersion.Contains(BetaUnityBuild))
+	            {
+	            }
+	            else
+	            {
+	               uScriptDebug.Log(BetaVersion + " (" + uScriptBuild + ") " + "will not work with Unity version " + Application.unityVersion + ".", uScriptDebug.Type.Error);
+	               return;
+	            }
+            
+			
+               if (DateTime.Now > ExpireDate)
+               {
+                  uScriptDebug.Log(BetaVersion + " (" + uScriptBuild + ") " + "has expired.\n", uScriptDebug.Type.Error);
+                  return;
+               }
+               else
+               {
+                  uScriptDebug.Log(BetaVersion + " (" + uScriptBuild + ") " + "will expire in " + (ExpireDate - DateTime.Now).Days + " days.", uScriptDebug.Type.Message);
+               }
+			 }
          }
 
          m_ScriptEditorCtrl = null;
